refactored some useful common code out of SecurityUtilsTest into AbstractSecurityTestBase
